changes the AddressFactory used for starting the directory scanner back to the default from world.
This is important since we are using the address type to decide whether an actor will start on a different node. The AddressFactory on Grid creates distributed addresses by default and therefore we must somehow make sure that it will not attempt to start/move the scanner on/to a different node.